sapju/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary
pydantic

python-dotenv
httpx
requests
pytest
pytest-cov

python-multipart
simple-file-checksum
prometheus-fastapi-instrumentator
